pyproject.toml
[tool.poetry]
name = "acquity"
version = "0.1.0"
description = ""
authors = ["Herbert Ilhan Tanujaya <mail@herbert.id>"]

[tool.poetry.dependencies]
python = "^3.7"
sanic = "^19.6"
sqlalchemy = "^1.3.9"
psycopg2 = "^2.8.3"
alembic = "^1.2.1"
sanic-jwt = "^1.3.2"
Sanic-Cors = "^0.9.9"
passlib = {version = "^1.7.1", extras = ["argon2"]}
cerberus = "^1.3.1"
python-dotenv = "^0.10.3"
[tool.poetry.dev-dependencies]
pytest = "^3.0"
black = {version = "^18.3-alpha.0", allows-prereleases = true}
isort = "^4.3.21"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
